feat: move module components inside ui layers
