+++ src/modules/posts/ui/components/PostCard.tsx
