+++ src/modules/posts/ui/components/PostsList.tsx
