+++ src/pages/users/index.tsx
@@ -2,7 +2,7 @@ import { User } from '@/src/modules/users/domain/User';
 import { UserRepository } from '@/src/modules/users/domain/UserRepository';
 import { createApiUserRepository } from '@/src/modules/users/infra/ApiUserRepository';
 import { getAllUsers } from '@/src/modules/users/application/get-all/getAllUsers';
-import { UsersList } from '@/src/sections/users/UsersList';
+import { UsersList } from '@/src/modules/users/ui/components/UsersList';
 
 interface UsersPageProps {
   users: User[];
